small bugs solved
quantity changed from int to float
checking existance before deleting

+++ finance/models.py
@@ -16,7 +16,7 @@ class Item(models.Model):
     budget = models.ForeignKey(Budget)
     department = models.ForeignKey(Department)
     name = models.CharField(max_length=50)
-    quantity=models.IntegerField(blank=True,null=True)
+    quantity=models.FloatField(blank=True,null=True)
     description = models.TextField(blank=True)
     comment = models.TextField(blank=True)
     original_amount = models.FloatField()
